+++ next/src/form.ts
@@ -4,7 +4,8 @@ import type { JsfObjectSchema, JsfSchema, SchemaValue } from './types'
 import type { ValidationOptions } from './validation/schema'
 import { getErrorMessage } from './errors/messages'
 import { buildFieldSchema } from './field/schema'
-import { applyComputedAttrsToSchema, mutateFields } from './mutations'
+import { mutateFields } from './mutations'
+import { applyComputedAttrsToSchema } from './validation/json-logic'
 import { validateSchema } from './validation/schema'
 
 export { ValidationOptions } from './validation/schema'
