+++ next/src/validation/json-logic.ts
@@ -217,3 +217,80 @@ export function computePropertyValues(
 
   return result
 }
+
+/**
+ * Applies any computed attributes to a schema, based on the provided values. When there are values to apply,
+ * it creates a deep clone of the schema and applies the computed values to the clone,otherwise it returns the original schema.
+ *
+ * @param schema - The schema to apply computed attributes to
+ * @param values - The current form values
+ * @returns The schema with computed attributes applied
+ */
+export function applyComputedAttrsToSchema(schema: JsfObjectSchema, values: SchemaValue) {
+  // If the schema has any computed attributes, we need to apply them by cloning the schema and applying the computed values
+  // Otherwise, we return the original schema
+  if (schema['x-jsf-logic']?.computedValues) {
+    // apply the computed values to the schema
+    const computedValuesDefinition = schema['x-jsf-logic'].computedValues
+    const computedValues: Record<string, string> = {}
+
+    Object.entries(computedValuesDefinition).forEach(([name, definition]) => {
+      const computedValue = computePropertyValues(definition.rule, values)
+      computedValues[name] = computedValue
+    })
+
+    const schemaCopy = safeDeepClone(schema)
+
+    cycleThroughPropertiesAndApplyValues(schemaCopy, computedValues)
+
+    return schemaCopy
+  }
+  else {
+    return schema
+  }
+}
+
+/**
+ * Cycles through the properties of a schema and applies the computed values to it
+ * @param schemaCopy - The schema to apply computed values to
+ * @param computedValues - The computed values to apply
+ */
+function cycleThroughPropertiesAndApplyValues(schemaCopy: JsfObjectSchema, computedValues: Record<string, string>) {
+  for (const propertyName in schemaCopy.properties) {
+    const computedAttrs = schemaCopy.properties[propertyName]['x-jsf-logic-computedAttrs']
+    if (computedAttrs) {
+      const propertySchema = schemaCopy.properties[propertyName] as JsfObjectSchema
+      cycleThroughAttrsAndApplyValues(propertySchema, computedValues, computedAttrs)
+
+      if (propertySchema.type === 'object' && propertySchema.properties) {
+        Object.entries(propertySchema.properties).forEach(([_, property]) => {
+          cycleThroughPropertiesAndApplyValues(property as JsfObjectSchema, computedValues)
+        })
+      }
+      delete propertySchema['x-jsf-logic-computedAttrs']
+    }
+  }
+}
+
+/**
+ * Cycles through the attributes of a schema and applies the computed values to it
+ * @param propertySchema - The schema to apply computed values to
+ * @param computedValues - The computed values to apply
+ */
+function cycleThroughAttrsAndApplyValues(propertySchema: JsfObjectSchema, computedValues: Record<string, string>, computedAttrs: JsfSchema['x-jsf-logic-computedAttrs']) {
+  for (const key in computedAttrs) {
+    const attributeName = key as keyof NonBooleanJsfSchema
+    const computationName = computedAttrs[key]
+    // const computedValue = computedValues[key]
+    // If the computation points to a string, it's a computed value and we can apply it directly
+    if (typeof computationName === 'string') {
+      const computedValue = computedValues[computationName]
+      propertySchema[attributeName] = computedValue
+    }
+    else {
+      Object.entries(computationName).forEach(([key, value]) => {
+        propertySchema[attributeName][key] = computedValues[value]
+      })
+    }
+  }
+}
